Check if git execution was successful before working with result
When the cryptopp-cmake directory was not checked out with Git, but
downloaded (i.e. from Github Releases), the extracted directory will
not be a git repository. Calling git in the directory will cause the
message 'fatal: not a git repository' during configuration, and the
following string(STRIP) operation will result in configuration failure.

This commit
- silences error output from Git to prevent users seeing the
  'fatal: ...' message
- checks the result code from Git to make sure the output variable is
  only used when Git was successful
- uses the GIT_CEILING_DIRECTORIES environment variable to prevent Git
  from ascending into parent directories of cryptopp-cmake and possibly
  finding a (different) git repository there.

Fixes abdes/cryptopp-cmake#94