Fix OpenMP on MacOS
Tested on Github Action macos-12, macos-13 (x64) and macos-14 (arm64) runners, they aren't able to find OpenMP. I don't know why the setting of `CMAKE_PREFIX_PATH` is not enough, but it does not work as intended. This change reinstates the code from the old cryptopp-cmake project, and I confirmed this works.